# fib_openapi3.yaml
openapi: 3.0.1
info:
  title: Fibonacci API
  description: >-
    An API to help the consumers to find values from the Fabonacci's sequence.
    
    Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net,
    #swagger](http://swagger.io/irc/).      For this sample, you can use the api
    key `special-key` to test the authorization     filters.
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: thiagodevel@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
externalDocs:
  description: Find out more about Swagger
  url: 'http://swagger.io'
servers:
  - url: 'https://localhost:8000/'
tags:
  - name: fib
    description: Fibonacci resources
    externalDocs:
      description: Find out more
      url: 'http://swagger.io'
paths:
  /fib:
    get:
      tags:
        - fib
      summary: retrive fibonacci value 
      description: Finds a value in the Fibonacci sequence for a given sequence position (N)
      operationId: RetrieveNthFibonacci
      parameters:
        - name: n
          in: query
          description: the position in the Fibonacci sequence to query
          required: true
          schema:
            type: integer
            minimum: 0
            maximum: 1474
      responses:
        '200':
          description: successful operation
          content:
            text/plain:
              schema:
                type: string
              example: "55"
        '400':
          description: Invalid status value
          content: {}
      security:
        - fib_auth:
            - 'read:fibonacci'
  /fibonacci:
    get:
      tags:
        - fib
      summary: retrive fibonacci data 
      description: Finds a fibonacci data from the Fibonacci sequence for a given sequence position (N)
      operationId: RetrieveNthFibonacciJson
      parameters:
        - name: n
          in: query
          description: the position in the Fibonacci sequence to query
          required: true
          schema:
            type: integer
            minimum: 0
            maximum: 1474
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FibonacciData'
        '400':
          description: Invalid status value
          content: {}
      security:
        - fib_auth:
            - 'read:fibonacci'            
components:
  schemas:
    FibonacciData:
      type: object
      properties:
        fibonacciNthPosition:
          type: integer
          description: the position provided by the consumer
          example: 10
        value:
          type: string
          description: the value in fibonacci sequences at the position
          example: "55"
      required:
      - fibonacciNthPosition
      - value
  securitySchemes:
    fib_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: 'http://petstore.swagger.io/oauth/dialog'
          scopes:
            'read:fibonacci': read fibonacci data
    api_key:
      type: apiKey
      name: api_key
      in: header